pyproject.toml
[project]
name = "redbank_orchestrator"
version = "0.1.0"
description = "RedBank multi-agent orchestrator that classifies user intent and routes queries to specialist agents (Knowledge, Banking Ops) via A2A protocol."
authors = [{ name = "Red Hat AI Engineering", email = "ai-engineering@redhat.com" }]
license = { text = "MIT" }
requires-python = ">=3.12,<=3.14"
dependencies = [
    "fastapi>=0.132.0",
    "uvicorn[standard]>=0.41.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph-prebuilt>=1.0.8",
    "python-dotenv>=1.2.2",
    "setuptools>=80.9.0,<82.0.0",
    "httpx>=0.28.0",
    "a2a-sdk[http-server]>=0.3.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.0.0",
]
tracing = [
    "mlflow[kubernetes]>=3.11.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"
